big-data-project/spark/requirements.txt
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
appnope==0.1.3
    # via ipython
asttokens==2.2.1
    # via stack-data
backcall==0.2.0
    # via ipython
decorator==5.1.1
    # via ipython
executing==1.2.0
    # via stack-data
ipython==8.7.0
jedi==0.18.2
matplotlib-inline==0.1.6
numpy
parso
pexpect
pickleshare
prompt-toolkit
ptyprocess
pure-eval
py4j
pyarrow
pygments
pyspark
python-dateutil
pytz
six
stack-data
traitlets
wcwidth
hdfs
findspark
confluent_kafka
influxdb
pymongo
python-dotenv
influxdb_client
